rename phylopic.py to phylo_pic.py
This is for compatability with case-insensitive file systems
