+++ src/cc_catalog_airflow/dags/provider_api_scripts/test_phylo_pic.py
@@ -5,7 +5,7 @@ import requests
 from unittest.mock import patch, MagicMock
 import pytest
 
-import phylopic as pp
+import phylo_pic as pp
 
 RESOURCES = os.path.join(
     os.path.abspath(os.path.dirname(__file__)), 'tests/resources/phylopic'
